Translate error messages for end user
The plug-in stores the error message based on the current user language. This PR stores the message on English and translates the message on the client side, so on GLPI instances with multilanguage user, the task log will show the error message based on the current user language instead of the language of the user that the error message has been generated.